depart from config.mk scheme
I used config.mk because i thought it was the easier way for users to
configure the building process to fit their system (and because that is
what suckless software uses).  But i then realised i should have instead
just use make variables for that.

Suckless is also moving away from config.mk in some projects[1].

[1]: https://git.suckless.org/tabbed/commit/41e2b8f30667078a4da04d5bd3d9eb839b118895.html